+++ packages/core/src/index.ts
@@ -31,5 +31,12 @@ export {
   OnAuthenticateLoginObject,
 } from './login';
 export { RevertableArraySorter } from './revertableArraySorter';
-export { sleepPromise, transformDateInRequest, apiUrl } from './utils';
+export {
+  sleepPromise,
+  transformDateInRequest,
+  apiUrl,
+  promiseAllAtOnce,
+  promiseAllWithData,
+  promiseEachInSequence,
+} from './utils';
 export { TestUtils, Constants, GraphUtils };